+++ R/misc_functions.R
@@ -1,6 +1,6 @@
 #' Logit
 #' @description Logit, defined as
-#' \deqn{log\left(\frac{x}{1-x}\right)}
+#' \eqn{log\left(\frac{x}{1-x}\right)}
 #' @param x Numeric vector
 #' @return Numeric vector
 #' @seealso \link{expit}
@@ -17,7 +17,7 @@ logit <- function(x) log(x/(1-x))
 #' Expit, or inverse logit
 #' @description Inverse logit, where logit is defined as \eqn{log\left(\frac{x}{1-x}\right)}.
 #'
-#' Expit (inverse logit) is defined as \deqn{\frac{exp(x)}{1+exp(x)}}
+#' Expit (inverse logit) is defined as \eqn{\frac{exp(x)}{1+exp(x)}}
 #' @param x Numeric vector
 #' @return Numeric vector
 #' @seealso \link{logit}
@@ -30,3 +30,24 @@ logit <- function(x) log(x/(1-x))
 #' @export
 expit <- function(x) exp(x)/(1+exp(x))
 
+
+
+#' Standard Error
+#' @description Standard error, defined as
+#' \eqn{\frac{\sigma_x}{\sqrt{n}}}
+#' @param x Numeric vector
+#' @param na.rm Logical.  Should missing values be removed?  Defaults to `FALSE`
+#' for consistency with \link{sd}.  Note: if `na.rm==TRUE`, the denominator will
+#' be the number of non-NA entries.
+#' @return Numeric of length 1
+#' @author Matt Tyers
+#' @examples
+#' a <- c(8, 6, 7, 5, 3, 0, 9, NA)
+#'
+#' se(a)
+#' se(a, na.rm=TRUE)
+#' sd(a, na.rm=TRUE)/sqrt(7)
+#' @export
+se <- function(x, na.rm=FALSE) {
+  sd(x, na.rm=na.rm)/sqrt(sum(!is.na(x)))
+}