+++ NEWS.md
@@ -3,4 +3,4 @@
 Initial creation.  Functions currently include
 
 * For ASL summaries: `ASL_table()`, `verify_ASL_table()`, and `ASL_boilerplate()`
-* Miscellaneous functions: `logit()` and `expit()`
+* Miscellaneous functions: `logit()`, `expit()`, and `se()`